Refactor updates module architecture
